TST:Fix all broken tests from moving things around
